+++ OrderHandlerMicroservice/Requests/GetOrderStatusRequest.cs
@@ -1,4 +1,4 @@
 ﻿namespace OrderHandlerMicroservice.Requests;
 
 public record GetOrderStatusRequest(
-    long OrderId);
+    int OrderId);
